Variable Explorer: Starting to move its widgets to its own namespace
