+++ spyderlib/widgets/externalshell/namespacebrowser.py
@@ -30,7 +30,7 @@ from spyderlib.utils.programs import is_module_installed
 from spyderlib.utils.qthelpers import (create_toolbutton, add_actions,
                                        create_action)
 from spyderlib.utils.iofuncs import iofunctions
-from spyderlib.widgets.importwizard import ImportWizard
+from spyderlib.widgets.varexp.importwizard import ImportWizard
 from spyderlib.config.base import _, get_supported_types
 from spyderlib.py3compat import is_text_string, to_text_string, getcwd
